+++ src/idc/imgstats/writer/__init__.py
@@ -1,2 +1,3 @@
 from ._area_histogram import AreaHistogramWriter
 from ._label_dist import LabelDistributionWriter
+from ._pixel_count import PixelCountWriter