+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - switched to `kasperl` library for base API and generic pipeline plugins
+- added `pixel-count` writer that outputs pixel counts per label for image segmentation annotations
 
 
 0.0.2 (2025-03-14)